+++ backend/src/main/java/shook/shook/member_part/domain/repository/dto/SongMemberPartCreatedAtDto.java
@@ -1,6 +1,6 @@
-package shook.shook.my_part.domain.repository.dto;
+package shook.shook.member_part.domain.repository.dto;
 
-import shook.shook.my_part.domain.MemberPart;
+import shook.shook.member_part.domain.MemberPart;
 import shook.shook.song.domain.Song;
 
 public interface SongMemberPartCreatedAtDto {
