+++ backend/src/main/java/shook/shook/song/application/dto/SongSwipeResponse.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import shook.shook.my_part.domain.MemberPart;
+import shook.shook.member_part.domain.MemberPart;
 import shook.shook.song.domain.Song;
 
 @Schema(description = "첫 스와이프 시, 현재 노래와 이전, 이후 노래 리스트 조회 응답")
