+++ backend/src/test/java/shook/shook/song/application/SongServiceTest.java
@@ -15,8 +15,8 @@ import shook.shook.auth.ui.Authority;
 import shook.shook.auth.ui.argumentresolver.MemberInfo;
 import shook.shook.member.domain.Member;
 import shook.shook.member.domain.repository.MemberRepository;
-import shook.shook.my_part.domain.MemberPart;
-import shook.shook.my_part.domain.repository.MemberPartRepository;
+import shook.shook.member_part.domain.MemberPart;
+import shook.shook.member_part.domain.repository.MemberPartRepository;
 import shook.shook.song.application.dto.KillingPartRegisterRequest;
 import shook.shook.song.application.dto.MemberPartResponse;
 import shook.shook.song.application.dto.SongResponse;
