+++ backend/src/test/java/shook/shook/song/ui/SongSwipeControllerTest.java
@@ -1,6 +1,9 @@
 package shook.shook.song.ui;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import io.restassured.RestAssured;
+import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -13,18 +16,14 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.jdbc.Sql;
 import shook.shook.auth.application.TokenProvider;
-import shook.shook.my_part.application.MemberPartService;
-import shook.shook.my_part.application.dto.MemberPartRegisterRequest;
+import shook.shook.member_part.application.MemberPartService;
+import shook.shook.member_part.application.dto.MemberPartRegisterRequest;
 import shook.shook.song.application.InMemorySongsScheduler;
 import shook.shook.song.application.dto.SongResponse;
 import shook.shook.song.application.dto.SongSwipeResponse;
 import shook.shook.song.application.killingpart.KillingPartLikeService;
 import shook.shook.song.application.killingpart.dto.HighLikedSongResponse;
 import shook.shook.song.application.killingpart.dto.KillingPartLikeRequest;
-
-import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 @Sql("classpath:/killingpart/initialize_killing_part_song.sql")
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
@@ -63,32 +62,32 @@ class SongSwipeControllerTest {
 
         final Long songId = 2L;
         likeService.updateLikeStatus(FIRST_SONG_KILLING_PART_ID_1, MEMBER_ID,
-                new KillingPartLikeRequest(true));
+                                     new KillingPartLikeRequest(true));
         likeService.updateLikeStatus(FIRST_SONG_KILLING_PART_ID_2, MEMBER_ID,
-                new KillingPartLikeRequest(true));
+                                     new KillingPartLikeRequest(true));
         likeService.updateLikeStatus(SECOND_SONG_KILLING_PART_ID_1, MEMBER_ID,
-                new KillingPartLikeRequest(true));
+                                     new KillingPartLikeRequest(true));
 
         inMemorySongsScheduler.recreateCachedSong();
         // 정렬 순서: 1L, 2L, 4L, 3L
         //when
         final SongSwipeResponse response = RestAssured.given().log().all()
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                .when().log().all()
-                .get("/songs/high-liked/{song_id}", songId)
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value())
-                .extract()
-                .body().as(SongSwipeResponse.class);
+            .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+            .when().log().all()
+            .get("/songs/high-liked/{song_id}", songId)
+            .then().log().all()
+            .statusCode(HttpStatus.OK.value())
+            .extract()
+            .body().as(SongSwipeResponse.class);
 
         //then
         assertThat(response.getPrevSongs().stream()
-                .map(SongResponse::getId).toList())
-                .containsExactly(1L);
+                       .map(SongResponse::getId).toList())
+            .containsExactly(1L);
         assertThat(response.getCurrentSong().getId()).isEqualTo(songId);
         assertThat(response.getNextSongs().stream()
-                .map(SongResponse::getId).toList())
-                .containsExactly(4L, 3L);
+                       .map(SongResponse::getId).toList())
+            .containsExactly(4L, 3L);
     }
 
     @DisplayName("가장 좋아요가 적은 노래 id 로 이전 노래 정보를 조회할 때 200 상태코드, 이전 노래 리스트를 반환한다.")
@@ -98,9 +97,9 @@ class SongSwipeControllerTest {
         final Long songId = 2L;
         final String accessToken = tokenProvider.createAccessToken(MEMBER_ID, "nickname");
         likeService.updateLikeStatus(FIRST_SONG_KILLING_PART_ID_1, MEMBER_ID,
-                new KillingPartLikeRequest(true));
+                                     new KillingPartLikeRequest(true));
         likeService.updateLikeStatus(FIRST_SONG_KILLING_PART_ID_2, MEMBER_ID,
-                new KillingPartLikeRequest(true));
+                                     new KillingPartLikeRequest(true));
 
         memberPartService.register(1L, MEMBER_ID, new MemberPartRegisterRequest(5, 5));
 
@@ -109,18 +108,18 @@ class SongSwipeControllerTest {
 
         //when
         final List<SongResponse> response = RestAssured.given().log().all()
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                .when().log().all()
-                .get("/songs/high-liked/{song_id}/prev", songId)
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value())
-                .extract()
-                .body().jsonPath().getList(".", SongResponse.class);
+            .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+            .when().log().all()
+            .get("/songs/high-liked/{song_id}/prev", songId)
+            .then().log().all()
+            .statusCode(HttpStatus.OK.value())
+            .extract()
+            .body().jsonPath().getList(".", SongResponse.class);
 
         // then
         assertThat(response.stream()
-                .map(SongResponse::getId).toList())
-                .containsExactly(1L, 4L, 3L);
+                       .map(SongResponse::getId).toList())
+            .containsExactly(1L, 4L, 3L);
         assertThat(response.get(0).getMemberPart().getPartId()).isNotNull();
         assertThat(response.get(1).getMemberPart()).isNull();
         assertThat(response.get(2).getMemberPart()).isNull();
@@ -139,18 +138,18 @@ class SongSwipeControllerTest {
 
         //when
         final List<SongResponse> response = RestAssured.given().log().all()
-                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                .when().log().all()
-                .get("/songs/high-liked/{song_id}/next", songId)
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value())
-                .extract()
-                .body().jsonPath().getList(".", SongResponse.class);
+            .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+            .when().log().all()
+            .get("/songs/high-liked/{song_id}/next", songId)
+            .then().log().all()
+            .statusCode(HttpStatus.OK.value())
+            .extract()
+            .body().jsonPath().getList(".", SongResponse.class);
 
         // then
         assertThat(response.stream()
-                .map(SongResponse::getId).toList())
-                .containsExactly(2L, 1L);
+                       .map(SongResponse::getId).toList())
+            .containsExactly(2L, 1L);
         assertThat(response.get(0).getMemberPart()).isNull();
         assertThat(response.get(1).getMemberPart()).isNotNull();
     }
@@ -171,19 +170,19 @@ class SongSwipeControllerTest {
 
             // when
             final List<HighLikedSongResponse> response = RestAssured.given().log().all()
-                    .queryParam("genre", genre)
-                    .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                    .when().log().all()
-                    .get("/songs/high-liked")
-                    .then().log().all()
-                    .statusCode(HttpStatus.OK.value())
-                    .extract()
-                    .body().jsonPath().getList(".", HighLikedSongResponse.class);
+                .queryParam("genre", genre)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+                .when().log().all()
+                .get("/songs/high-liked")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract()
+                .body().jsonPath().getList(".", HighLikedSongResponse.class);
 
             // then
             assertThat(response.stream()
-                    .map(HighLikedSongResponse::getId).toList())
-                    .containsExactly(4L, 3L, 1L);
+                           .map(HighLikedSongResponse::getId).toList())
+                .containsExactly(4L, 3L, 1L);
         }
 
         @DisplayName("존재하지 않는 장르를 요청한 경우 400 상태코드를 반환한다.")
@@ -196,14 +195,14 @@ class SongSwipeControllerTest {
             // when
             // then
             RestAssured.given().log().all()
-                    .queryParam("genre", genre)
-                    .param("memberId", MEMBER_ID)
-                    .when().log().all()
-                    .get("/songs/high-liked")
-                    .then().log().all()
-                    .statusCode(HttpStatus.BAD_REQUEST.value())
-                    .extract()
-                    .body().jsonPath().getString("genre");
+                .queryParam("genre", genre)
+                .param("memberId", MEMBER_ID)
+                .when().log().all()
+                .get("/songs/high-liked")
+                .then().log().all()
+                .statusCode(HttpStatus.BAD_REQUEST.value())
+                .extract()
+                .body().jsonPath().getString("genre");
         }
 
         @DisplayName("비회원이 장르 노래 정보를 처음으로 조회할 때, 가운데 장르 노래를 기준으로 조회한 경우 200 상태코드, 현재 노래, 이전 / 이후 장르 노래 리스트를 반환한다.")
@@ -215,11 +214,11 @@ class SongSwipeControllerTest {
             final String accessToken = tokenProvider.createAccessToken(MEMBER_ID, "nickname");
 
             likeService.updateLikeStatus(FIRST_SONG_KILLING_PART_ID_1, MEMBER_ID,
-                    new KillingPartLikeRequest(true));
+                                         new KillingPartLikeRequest(true));
             likeService.updateLikeStatus(FIRST_SONG_KILLING_PART_ID_2, MEMBER_ID,
-                    new KillingPartLikeRequest(true));
+                                         new KillingPartLikeRequest(true));
             likeService.updateLikeStatus(SECOND_SONG_KILLING_PART_ID_1, MEMBER_ID,
-                    new KillingPartLikeRequest(true));
+                                         new KillingPartLikeRequest(true));
 
             memberPartService.register(1L, MEMBER_ID, new MemberPartRegisterRequest(5, 5));
 
@@ -228,23 +227,23 @@ class SongSwipeControllerTest {
 
             //when
             final SongSwipeResponse response = RestAssured.given().log().all()
-                    .queryParam("genre", genre)
-                    .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                    .when().log().all()
-                    .get("/songs/high-liked/{songId}", songId)
-                    .then().log().all()
-                    .statusCode(HttpStatus.OK.value())
-                    .extract()
-                    .body().as(SongSwipeResponse.class);
+                .queryParam("genre", genre)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+                .when().log().all()
+                .get("/songs/high-liked/{songId}", songId)
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract()
+                .body().as(SongSwipeResponse.class);
 
             //then
             assertThat(response.getPrevSongs().stream()
-                    .map(SongResponse::getId).toList())
-                    .containsExactly(1L);
+                           .map(SongResponse::getId).toList())
+                .containsExactly(1L);
             assertThat(response.getCurrentSong().getId()).isEqualTo(songId);
             assertThat(response.getNextSongs().stream()
-                    .map(SongResponse::getId).toList())
-                    .containsExactly(3L);
+                           .map(SongResponse::getId).toList())
+                .containsExactly(3L);
             assertThat(response.getPrevSongs().get(0).getMemberPart()).isNotNull();
             assertThat(response.getCurrentSong().getMemberPart()).isNull();
             assertThat(response.getNextSongs().get(0).getMemberPart()).isNull();
@@ -259,9 +258,9 @@ class SongSwipeControllerTest {
             final String accessToken = tokenProvider.createAccessToken(MEMBER_ID, "nickname");
 
             likeService.updateLikeStatus(FIRST_SONG_KILLING_PART_ID_1, MEMBER_ID,
-                    new KillingPartLikeRequest(true));
+                                         new KillingPartLikeRequest(true));
             likeService.updateLikeStatus(FIRST_SONG_KILLING_PART_ID_2, MEMBER_ID,
-                    new KillingPartLikeRequest(true));
+                                         new KillingPartLikeRequest(true));
 
             memberPartService.register(1L, MEMBER_ID, new MemberPartRegisterRequest(5, 5));
 
@@ -270,19 +269,19 @@ class SongSwipeControllerTest {
 
             //when
             final List<SongResponse> response = RestAssured.given().log().all()
-                    .queryParam("genre", genre)
-                    .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                    .when().log().all()
-                    .get("/songs/high-liked/{songId}/prev", songId)
-                    .then().log().all()
-                    .statusCode(HttpStatus.OK.value())
-                    .extract()
-                    .body().jsonPath().getList(".", SongResponse.class);
+                .queryParam("genre", genre)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+                .when().log().all()
+                .get("/songs/high-liked/{songId}/prev", songId)
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract()
+                .body().jsonPath().getList(".", SongResponse.class);
 
             //then
             assertThat(response.stream()
-                    .map(SongResponse::getId).toList())
-                    .containsExactly(1L, 4L);
+                           .map(SongResponse::getId).toList())
+                .containsExactly(1L, 4L);
             assertThat(response.get(0).getMemberPart()).isNotNull();
             assertThat(response.get(1).getMemberPart()).isNull();
         }
@@ -296,11 +295,11 @@ class SongSwipeControllerTest {
             final String accessToken = tokenProvider.createAccessToken(MEMBER_ID, "nickname");
 
             likeService.updateLikeStatus(FIRST_SONG_KILLING_PART_ID_1, MEMBER_ID,
-                    new KillingPartLikeRequest(true));
+                                         new KillingPartLikeRequest(true));
             likeService.updateLikeStatus(FIRST_SONG_KILLING_PART_ID_2, MEMBER_ID,
-                    new KillingPartLikeRequest(true));
+                                         new KillingPartLikeRequest(true));
             likeService.updateLikeStatus(SECOND_SONG_KILLING_PART_ID_1, MEMBER_ID,
-                    new KillingPartLikeRequest(true));
+                                         new KillingPartLikeRequest(true));
 
             memberPartService.register(4L, MEMBER_ID, new MemberPartRegisterRequest(5, 5));
 
@@ -309,19 +308,19 @@ class SongSwipeControllerTest {
 
             //when
             final List<SongResponse> response = RestAssured.given().log().all()
-                    .queryParam("genre", genre)
-                    .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
-                    .when().log().all()
-                    .get("/songs/high-liked/{songId}/next", songId)
-                    .then().log().all()
-                    .statusCode(HttpStatus.OK.value())
-                    .extract()
-                    .body().jsonPath().getList(".", SongResponse.class);
+                .queryParam("genre", genre)
+                .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+                .when().log().all()
+                .get("/songs/high-liked/{songId}/next", songId)
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract()
+                .body().jsonPath().getList(".", SongResponse.class);
 
             //then
             assertThat(response.stream()
-                    .map(SongResponse::getId).toList())
-                    .containsExactly(4L, 3L);
+                           .map(SongResponse::getId).toList())
+                .containsExactly(4L, 3L);
             assertThat(response.get(0).getMemberPart()).isNotNull();
             assertThat(response.get(1).getMemberPart()).isNull();
         }
